cofirm delete

+++ client/pages/okr/okr.js
@@ -15,10 +15,10 @@ Page({
     let id = event.currentTarget.dataset.id;
     let status = event.currentTarget.dataset.status;
     let statusChange = status ? 0 : 1;
-    let statusChangeDisplay = statusChange ? '标记为已完成' : '标记为未完成';
+    let statusChangeDisplay = statusChange ? '已完成' : '未完成';
 
     wx.showActionSheet({
-      itemList: [ '查看 OKR 详情',statusChangeDisplay, '删除'],
+      itemList: [ '查看', '编辑',statusChangeDisplay, '删除'],
       itemColor: '#333',
       success: (res)=> {
         let tapIndex = res.tapIndex;
@@ -27,10 +27,21 @@ Page({
             wx.navigateTo({url:'/pages/okr_detail/okr_detail?id=' + id })
             break;
           case 1:
-            this.handleChangeObjective(id,index,statusChange)
+            wx.navigateTo({url:'/pages/okr_edit/okr_edit?id=' + id })
             break;
           case 2:
-            this.handleDeleteObjective(id,index)
+            this.handleChangeObjective(id,index,statusChange)
+            break;
+          case 3:
+            wx.showModal({
+              title: '删除',
+              content: '是否确定要删除该 OKR',
+              success:(res)=> {
+                if (res.confirm) {
+                  this.handleDeleteObjective(id,index)
+                }
+              }
+            })
             break;
         }
       },
